cleanup: refactor logging helpers.
We are going to need these helpers to log from DatabaseAdminClient, and
eventually from InstanceAdminClient too.  This change just refactors
them to a separate file.
